Change memory allocation for blocks
As soon as a root block is ready (i.e. no more blocks will be added),
previous root blocks are already send to parse_spans. When these
previous blocks are send, then the memory associated will be free for
reuse. This means that memory allocation go only as the deepest nested
block of the whole document.